Completed splitting server into multiple files
Deleted unnecessary files
